removed packages/app and backend
