--- workspaces/topology/packages/app-legacy/.eslintignore
@@ -1,1 +1,0 @@
-public
