--- workspaces/topology/packages/app-legacy/knip-report.md
@@ -1,8 +1,0 @@
-# Knip report
-
-## Unused dependencies (1)
-
-| Name         | Location          | Severity |
-| :----------- | :---------------- | :------- |
-| react-router | package.json:51:6 | error    |
-
