--- workspaces/topology/packages/app-legacy/package.json
@@ -1,76 +1,0 @@
-{
-  "name": "app-legacy",
-  "version": "0.0.0",
-  "private": true,
-  "bundled": true,
-  "backstage": {
-    "role": "frontend"
-  },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/backstage/community-plugins",
-    "directory": "workspaces/topology/packages/app-legacy"
-  },
-  "scripts": {
-    "start": "backstage-cli package start",
-    "build": "backstage-cli package build",
-    "clean": "backstage-cli package clean",
-    "test": "backstage-cli package test",
-    "lint": "backstage-cli package lint"
-  },
-  "dependencies": {
-    "@backstage-community/plugin-topology": "workspace:^",
-    "@backstage/app-defaults": "^1.7.6",
-    "@backstage/catalog-model": "^1.7.7",
-    "@backstage/cli": "^0.36.0",
-    "@backstage/core-app-api": "^1.19.6",
-    "@backstage/core-components": "^0.18.8",
-    "@backstage/core-plugin-api": "^1.12.4",
-    "@backstage/integration-react": "^1.2.16",
-    "@backstage/plugin-api-docs": "^0.13.5",
-    "@backstage/plugin-catalog": "^2.0.1",
-    "@backstage/plugin-catalog-common": "^1.1.8",
-    "@backstage/plugin-catalog-graph": "^0.6.0",
-    "@backstage/plugin-catalog-import": "^0.13.11",
-    "@backstage/plugin-catalog-react": "^2.1.0",
-    "@backstage/plugin-kubernetes": "^0.12.17",
-    "@backstage/plugin-org": "^0.7.0",
-    "@backstage/plugin-permission-react": "^0.4.41",
-    "@backstage/plugin-scaffolder": "^1.36.1",
-    "@backstage/plugin-search": "^1.7.0",
-    "@backstage/plugin-search-react": "^1.11.0",
-    "@backstage/plugin-techdocs": "^1.17.2",
-    "@backstage/plugin-techdocs-module-addons-contrib": "^1.1.34",
-    "@backstage/plugin-techdocs-react": "^1.3.9",
-    "@backstage/plugin-user-settings": "^0.9.1",
-    "@backstage/theme": "^0.7.2",
-    "@backstage/ui": "^0.13.1",
-    "@material-ui/core": "^4.12.2",
-    "@material-ui/icons": "^4.9.1",
-    "react": "^18.0.2",
-    "react-dom": "^18.0.2",
-    "react-router": "^6.30.2",
-    "react-router-dom": "^6.30.2"
-  },
-  "devDependencies": {
-    "@playwright/test": "^1.32.3",
-    "@testing-library/jest-dom": "^6.0.0",
-    "@testing-library/react": "^16.0.0",
-    "@types/react-dom": "*"
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
-  },
-  "files": [
-    "dist"
-  ]
-}
