--- workspaces/topology/packages/app-legacy/public/index.html
@@ -1,60 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="utf-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1" />
-    <meta name="theme-color" content="#000000" />
-    <meta
-      name="description"
-      content="Backstage is an open source framework for building developer portals"
-    />
-    <!--
-      manifest.json provides metadata used when your web app is installed on a
-      user's mobile device or desktop. See https://developers.google.com/web/fundamentals/web-app-manifest/
-    -->
-    <link
-      rel="manifest"
-      href="<%= publicPath %>/manifest.json"
-      crossorigin="use-credentials"
-    />
-    <link rel="icon" href="<%= publicPath %>/favicon.ico" />
-    <link rel="shortcut icon" href="<%= publicPath %>/favicon.ico" />
-    <link
-      rel="apple-touch-icon"
-      sizes="180x180"
-      href="<%= publicPath %>/apple-touch-icon.png"
-    />
-    <link
-      rel="icon"
-      type="image/png"
-      sizes="32x32"
-      href="<%= publicPath %>/favicon-32x32.png"
-    />
-    <link
-      rel="icon"
-      type="image/png"
-      sizes="16x16"
-      href="<%= publicPath %>/favicon-16x16.png"
-    />
-    <link
-      rel="mask-icon"
-      href="<%= publicPath %>/safari-pinned-tab.svg"
-      color="#5bbad5"
-    />
-    <title><%= config.getOptionalString('app.title') ?? 'Backstage' %></title>
-  </head>
-  <body>
-    <noscript>You need to enable JavaScript to run this app.</noscript>
-    <div id="root"></div>
-    <!--
-      This HTML file is a template.
-      If you open it directly in the browser, you will see an empty page.
-
-      You can add webfonts, meta tags, or analytics to this file.
-      The build step will place the bundled scripts into the <body> tag.
-
-      To begin the development, run `yarn start`.
-      To create a production bundle, use `yarn build`.
-    -->
-  </body>
-</html>
