--- workspaces/topology/packages/app-legacy/public/manifest.json
@@ -1,15 +1,0 @@
-{
-  "short_name": "Backstage",
-  "name": "Backstage",
-  "icons": [
-    {
-      "src": "favicon.ico",
-      "sizes": "48x48",
-      "type": "image/png"
-    }
-  ],
-  "start_url": "./index.html",
-  "display": "standalone",
-  "theme_color": "#000000",
-  "background_color": "#ffffff"
-}
