--- workspaces/topology/packages/app-legacy/src/components/search/SearchPage.tsx
@@ -1,123 +1,0 @@
-/*
- * Copyright 2024 The Backstage Authors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-import { Card, Grid } from '@backstage/ui';
-
-import { CatalogSearchResultListItem } from '@backstage/plugin-catalog';
-import {
-  catalogApiRef,
-  CATALOG_FILTER_EXISTS,
-} from '@backstage/plugin-catalog-react';
-import { TechDocsSearchResultListItem } from '@backstage/plugin-techdocs';
-
-import { SearchType } from '@backstage/plugin-search';
-import {
-  SearchBar,
-  SearchFilter,
-  SearchResult,
-  SearchPagination,
-  useSearch,
-} from '@backstage/plugin-search-react';
-import {
-  CatalogIcon,
-  Content,
-  DocsIcon,
-  Header,
-  Page,
-} from '@backstage/core-components';
-import { useApi } from '@backstage/core-plugin-api';
-
-import styles from './SearchPage.module.css';
-
-const SearchPage = () => {
-  const { types } = useSearch();
-  const catalogApi = useApi(catalogApiRef);
-
-  return (
-    <Page themeId="home">
-      <Header title="Search" />
-      <Content>
-        <Grid.Root columns={{ sm: '12' }} gap="6">
-          <Grid.Item colSpan={{ sm: '12' }}>
-            <Card className={styles.bar}>
-              <SearchBar />
-            </Card>
-          </Grid.Item>
-          <Grid.Item colSpan={{ sm: '3' }}>
-            <SearchType.Accordion
-              name="Result Type"
-              defaultValue="software-catalog"
-              types={[
-                {
-                  value: 'software-catalog',
-                  name: 'Software Catalog',
-                  icon: <CatalogIcon />,
-                },
-                {
-                  value: 'techdocs',
-                  name: 'Documentation',
-                  icon: <DocsIcon />,
-                },
-              ]}
-            />
-            <div className={styles.filters}>
-              {types.includes('techdocs') && (
-                <SearchFilter.Select
-                  className={styles.filter}
-                  label="Entity"
-                  name="name"
-                  values={async () => {
-                    const { items } = await catalogApi.getEntities({
-                      fields: ['metadata.name'],
-                      filter: {
-                        'metadata.annotations.backstage.io/techdocs-ref':
-                          CATALOG_FILTER_EXISTS,
-                      },
-                    });
-
-                    const names = items.map(entity => entity.metadata.name);
-                    names.sort();
-                    return names;
-                  }}
-                />
-              )}
-              <SearchFilter.Select
-                className={styles.filter}
-                label="Kind"
-                name="kind"
-                values={['Component', 'Template']}
-              />
-              <SearchFilter.Checkbox
-                className={styles.filter}
-                label="Lifecycle"
-                name="lifecycle"
-                values={['experimental', 'production']}
-              />
-            </div>
-          </Grid.Item>
-          <Grid.Item colSpan={{ sm: '9' }}>
-            <SearchPagination />
-            <SearchResult>
-              <CatalogSearchResultListItem icon={<CatalogIcon />} />
-              <TechDocsSearchResultListItem icon={<DocsIcon />} />
-            </SearchResult>
-          </Grid.Item>
-        </Grid.Root>
-      </Content>
-    </Page>
-  );
-};
-
-export const searchPage = <SearchPage />;
