--- workspaces/topology/packages/app/.eslintrc.js
@@ -1,1 +1,0 @@
-module.exports = require('@backstage/cli/config/eslint-factory')(__dirname);
