--- workspaces/topology/packages/app/knip-report.md
@@ -1,12 +1,0 @@
-# Knip report
-
-## Unused dependencies (5)
-
-| Name                         | Location          | Severity |
-| :--------------------------- | :---------------- | :------- |
-| @backstage/plugin-kubernetes | package.json:31:6 | error    |
-| @backstage/plugin-scaffolder | package.json:33:6 | error    |
-| @backstage/plugin-api-docs   | package.json:28:6 | error    |
-| @backstage/plugin-techdocs   | package.json:35:6 | error    |
-| @backstage/plugin-catalog    | package.json:30:6 | error    |
-
