--- workspaces/topology/packages/app/src/App.tsx
@@ -1,35 +1,0 @@
-/*
- * Copyright 2024 The Backstage Authors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-import { createApp } from '@backstage/frontend-defaults';
-import { rhdhThemeModule } from '@red-hat-developer-hub/backstage-plugin-theme/alpha';
-import {
-  topologyCatalogModule,
-  topologyTranslationsModule,
-} from '@backstage-community/plugin-topology/alpha';
-import { signInModule } from './modules/signIn';
-import { navModule } from './modules/nav';
-
-const app = createApp({
-  features: [
-    rhdhThemeModule,
-    topologyCatalogModule,
-    topologyTranslationsModule,
-    signInModule,
-    navModule,
-  ],
-});
-
-export default app.createRoot();
