--- workspaces/topology/packages/app/src/modules/nav/Sidebar.tsx
@@ -1,67 +1,0 @@
-/*
- * Copyright 2024 The Backstage Authors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-import {
-  Sidebar,
-  SidebarDivider,
-  SidebarGroup,
-  SidebarItem,
-  SidebarSpace,
-} from '@backstage/core-components';
-import { NavContentBlueprint } from '@backstage/plugin-app-react';
-import { SidebarSearchModal } from '@backstage/plugin-search';
-import { SidebarLogo } from './SidebarLogo';
-import { MyGroupsSidebarItem } from '@backstage/plugin-org';
-import HomeIcon from '@material-ui/icons/Home';
-import CreateIcon from '@material-ui/icons/AddCircleOutline';
-import ExtensionIcon from '@material-ui/icons/Extension';
-import GroupIcon from '@material-ui/icons/People';
-import LibraryBooksIcon from '@material-ui/icons/LibraryBooks';
-import MenuIcon from '@material-ui/icons/Menu';
-import SearchIcon from '@material-ui/icons/Search';
-import {
-  UserSettingsSignInAvatar,
-  Settings as SidebarSettings,
-} from '@backstage/plugin-user-settings';
-
-export const SidebarContent = NavContentBlueprint.make({
-  params: {
-    component: () => (
-      <Sidebar>
-        <SidebarLogo />
-        <SidebarGroup label="Search" icon={<SearchIcon />}>
-          <SidebarSearchModal />
-        </SidebarGroup>
-        <SidebarDivider />
-        <SidebarGroup label="Menu" icon={<MenuIcon />}>
-          <SidebarItem icon={HomeIcon} to="catalog" text="Catalog" />
-          <MyGroupsSidebarItem
-            singularTitle="My Group"
-            pluralTitle="My Groups"
-            icon={GroupIcon}
-          />
-          <SidebarItem icon={ExtensionIcon} to="api-docs" text="APIs" />
-          <SidebarItem icon={LibraryBooksIcon} to="docs" text="TechDocs" />
-          <SidebarItem icon={CreateIcon} to="create" text="Create..." />
-        </SidebarGroup>
-        <SidebarSpace />
-        <SidebarDivider />
-        <SidebarGroup label="Settings" icon={<UserSettingsSignInAvatar />}>
-          <SidebarSettings />
-        </SidebarGroup>
-      </Sidebar>
-    ),
-  },
-});
