--- workspaces/topology/packages/app/src/modules/signIn/SignInPageComponent.tsx
@@ -1,29 +1,0 @@
-/*
- * Copyright 2024 The Backstage Authors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-import { SignInPage } from '@backstage/core-components';
-import { githubAuthApiRef } from '@backstage/core-plugin-api';
-import type { SignInPageProps } from '@backstage/plugin-app-react';
-
-const githubProvider = {
-  id: 'github-auth-provider',
-  title: 'GitHub',
-  message: 'Sign in using GitHub',
-  apiRef: githubAuthApiRef,
-};
-
-export function SignInPageComponent(props: SignInPageProps) {
-  return <SignInPage {...props} auto providers={['guest', githubProvider]} />;
-}
