--- workspaces/topology/packages/app/src/modules/signIn/index.ts
@@ -1,29 +1,0 @@
-/*
- * Copyright 2024 The Backstage Authors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-import { createFrontendModule } from '@backstage/frontend-plugin-api';
-import { SignInPageBlueprint } from '@backstage/plugin-app-react';
-
-export const signInModule = createFrontendModule({
-  pluginId: 'app',
-  extensions: [
-    SignInPageBlueprint.make({
-      params: {
-        loader: () =>
-          import('./SignInPageComponent').then(m => m.SignInPageComponent),
-      },
-    }),
-  ],
-});
