--- workspaces/topology/packages/backend/knip-report.md
@@ -1,16 +1,0 @@
-# Knip report
-
-## Unused dependencies (9)
-
-| Name                                  | Location          | Severity |
-| :------------------------------------ | :---------------- | :------- |
-| @backstage/plugin-search-backend-node | package.json:45:6 | error    |
-| @backstage/plugin-permission-common   | package.json:37:6 | error    |
-| @backstage/plugin-permission-node     | package.json:38:6 | error    |
-| @backstage/plugin-auth-node           | package.json:30:6 | error    |
-| @backstage/config                     | package.json:25:6 | error    |
-| app-legacy                            | package.json:48:6 | error    |
-| node-gyp                              | package.json:50:6 | error    |
-| app                                   | package.json:47:6 | error    |
-| pg                                    | package.json:51:6 | error    |
-
