+++ workspaces/topology/plugins/topology/package.json
@@ -41,7 +41,7 @@
     "postpack": "backstage-cli package postpack",
     "prepack": "backstage-cli package prepack",
     "start": "backstage-cli package start",
-    "start:legacy": "backstage-cli package start --entrypoint dev/legacy.tsx",
+    "start:alpha": "backstage-cli package start --entrypoint dev/alpha.tsx",
     "test": "backstage-cli package test --passWithNoTests --coverage",
     "tsc": "tsc"
   },
@@ -60,14 +60,25 @@
     "@backstage/frontend-plugin-api": "^0.15.0",
     "@backstage/plugin-app-react": "^0.2.1",
     "@backstage/plugin-catalog": "^2.0.1",
+    "@backstage/core-compat-api": "^0.5.9",
+    "@backstage/errors": "^1.2.7",
+    "@backstage/filter-predicates": "^0.1.1",
+    "@backstage/plugin-kubernetes": "^0.12.17",
     "@backstage/plugin-user-settings": "^0.9.1",
     "@kubernetes/client-node": "1.4.0",
+    "@material-ui/core": "^4.12.4",
+    "@material-ui/icons": "^4.9.1",
+    "@mui/icons-material": "5.18.0",
+    "@mui/lab": "5.0.0-alpha.177",
+    "@mui/material": "^5.15.17",
+    "@mui/styles": "5.18.0",
     "@patternfly/patternfly": "^6.1.0",
     "@patternfly/react-charts": "^8.1.0",
     "@patternfly/react-core": "^6.1.0",
     "@patternfly/react-tokens": "^6.1.0",
     "@patternfly/react-topology": "^6.1.0",
     "@remixicon/react": "^4.2.0",
+    "@types/react": "^16.13.1 || ^17.0.0 || ^18.0.0",
     "classnames": "2.x",
     "git-url-parse": "^13.1.0",
     "js-yaml": "^4.1.0",
@@ -90,12 +101,14 @@
     "@playwright/test": "1.60.0",
     "@backstage/catalog-client": "^1.14.0",
     "@backstage/frontend-defaults": "^0.3.0",
+    "@backstage/plugin-catalog": "^2.0.1",
     "@backstage/ui": "^0.13.2",
     "@testing-library/dom": "^10.0.0",
     "@testing-library/jest-dom": "^6.0.0",
     "@testing-library/react": "^16.0.0",
     "@types/git-url-parse": "9.0.3",
     "@types/react": "^18.3.11",
+    "@types/react-dom": "^18",
     "cross-fetch": "4.1.0",
     "glob": "^13.0.0",
     "react": "^18.3.1",
