+++ workspaces/topology/tsconfig.json
@@ -1,11 +1,6 @@
 {
   "extends": "@backstage/cli/config/tsconfig.json",
-  "include": [
-    "packages/*/src",
-    "plugins/*/src",
-    "plugins/*/dev",
-    "plugins/*/migrations"
-  ],
+  "include": ["plugins/*/src", "plugins/*/dev", "plugins/*/migrations"],
   "files": ["node_modules/@backstage/cli/asset-types/asset-types.d.ts"],
   "exclude": ["node_modules"],
   "compilerOptions": {
